refactor: Game and GameLogic.
