+++ model/src/main/java/io/deeplay/camp/Main.java
@@ -1,11 +1,8 @@
 package io.deeplay.camp;
 
-import game.Game;
-
 //TIP To <b>Run</b> code, press <shortcut actionId="Run"/> or
 // click the <icon src="AllIcons.Actions.Execute"/> icon in the gutter.
 public class Main {
     public static void main(String[] args) {
-        new Game();
     }
 }
