+++ services/build.gradle.kts
@@ -10,6 +10,7 @@ repositories {
 }
 
 dependencies {
+    implementation(project(":model"))
     testImplementation(platform("org.junit:junit-bom:5.10.0"))
     testImplementation("org.junit.jupiter:junit-jupiter")
     implementation("ch.qos.logback:logback-classic:1.4.12")
