+++ services/src/main/java/repository/ReversiListener.java
@@ -1,4 +1,4 @@
-package listener;
+package repository;
 
 import enums.GameStatus;
